+++ models/attachments.js
@@ -1,5 +1,7 @@
 import { Meteor } from 'meteor/meteor';
 import { FilesCollection } from 'meteor/ostrio:files';
+import { exec } from 'node:child_process';
+import { promisify } from 'node:util';
 import { createBucket } from './lib/grid/createBucket';
 import fs from 'fs';
 import FileType from 'file-type';
@@ -7,12 +9,15 @@ import path from 'path';
 import { AttachmentStoreStrategyFilesystem, AttachmentStoreStrategyGridFs} from '/models/lib/attachmentStoreStrategy';
 import FileStoreStrategyFactory, {moveToStorage, rename, STORAGE_NAME_FILESYSTEM, STORAGE_NAME_GRIDFS} from '/models/lib/fileStoreStrategy';
 
+let asyncExec;
+let attachmentUploadExternalProgram;
 let attachmentUploadMimeTypes = [];
 let attachmentUploadSize = 0;
 let attachmentBucket;
 let storagePath;
 
 if (Meteor.isServer) {
+  asyncExec = promisify(exec);
   attachmentBucket = createBucket('attachments');
 
   if (process.env.ATTACHMENTS_UPLOAD_MIME_TYPES) {
@@ -28,6 +33,14 @@ if (Meteor.isServer) {
     }
   }
 
+  if (process.env.ATTACHMENTS_UPLOAD_EXTERNAL_PROGRAM) {
+    attachmentUploadExternalProgram = process.env.ATTACHMENTS_UPLOAD_EXTERNAL_PROGRAM;
+
+    if (!attachmentUploadExternalProgram.includes("{file}")) {
+      attachmentUploadExternalProgram = undefined;
+    }
+  }
+
   storagePath = path.join(process.env.WRITABLE_PATH, 'attachments');
 }
 
@@ -40,42 +53,33 @@ Attachments = new FilesCollection({
   debug: false, // Change to `true` for debugging
   collectionName: 'attachments',
   allowClientCode: true,
-/* Commenting out because this custom namingFunction did not work:
-   https://github.com/veliovgroup/Meteor-Files/issues/847
-
   namingFunction(opts) {
-    const filenameWithoutExtension = opts.meta.name.replace(/(.+)\..+/, "$1");
-    const ret = opts.meta.fileId + "-original-" + filenameWithoutExtension;
+    let filenameWithoutExtension = ""
+    let fileId = "";
+    if (opts?.name) {
+      // Client
+      filenameWithoutExtension = opts.name.replace(/(.+)\..+/, "$1");
+      fileId = opts.meta.fileId;
+      delete opts.meta.fileId;
+    } else if (opts?.file?.name) {
+      // Server
+      filenameWithoutExtension = opts.file.name.replace(new RegExp(opts.file.extensionWithDot + "$"), "")
+      fileId = opts.fileId;
+    }
+    else {
+      // should never reach here
+      filenameWithoutExtension = Math.random().toString(36).slice(2);
+      fileId = Math.random().toString(36).slice(2);
+    }
+    const ret = fileId + "-original-" + filenameWithoutExtension;
     // remove fileId from meta, it was only stored there to have this information here in the namingFunction function
-    delete opts.meta.fileId;
     return ret;
   },
-*/
   storagePath() {
     const ret = fileStoreStrategyFactory.storagePath;
     return ret;
   },
   onAfterUpload(fileObj) {
-    let isValid = true;
-
-    if (attachmentUploadMimeTypes.length) {
-      const mimeTypeResult = Promise.await(FileType.fromFile(fileObj.path));
-
-      const mimeType = (mimeTypeResult ? mimeTypeResult.mime : fileObj.type);
-      const baseMimeType = mimeType.split('/', 1)[0];
-
-      isValid = attachmentUploadMimeTypes.includes(mimeType) || attachmentUploadMimeTypes.includes(baseMimeType + '/*') || attachmentUploadMimeTypes.includes('*');
-
-      if (!isValid) {
-        console.log("Validation of uploaded file failed: file " + fileObj.path + " - mimetype " + mimeType);
-      }
-    }
-
-    if (attachmentUploadSize && fileObj.size > attachmentUploadSize) {
-      console.log("Validation of uploaded file failed: file " + fileObj.path + " - size " + fileObj.size);
-      isValid = false;
-    }
-
     // current storage is the filesystem, update object and database
     Object.keys(fileObj.versions).forEach(versionName => {
       fileObj.versions[versionName].storage = STORAGE_NAME_FILESYSTEM;
@@ -83,12 +87,8 @@ Attachments = new FilesCollection({
 
     Attachments.update({ _id: fileObj._id }, { $set: { "versions" : fileObj.versions } });
 
-    if (isValid) {
-      let storage = fileObj.meta.copyStorage || STORAGE_NAME_GRIDFS;
-      moveToStorage(fileObj, storage, fileStoreStrategyFactory);
-    } else {
-      this.remove(fileObj._id);
-    }
+    let storageDestination = fileObj.meta.copyStorage || STORAGE_NAME_GRIDFS;
+    Meteor.defer(() => Meteor.call('validateAttachmentAndMoveToStorage', fileObj._id, storageDestination));
   },
   interceptDownload(http, fileObj, versionName) {
     const ret = fileStoreStrategyFactory.getFileStrategy(fileObj, versionName).interceptDownload(http, this.cacheControl);
@@ -140,6 +140,9 @@ if (Meteor.isServer) {
 
       const fileObj = Attachments.findOne({_id: fileObjId});
       moveToStorage(fileObj, storageDestination, fileStoreStrategyFactory);
+
+      // since Meteor-Files 2.1.0 the filename is truncated to 28 characters, so rename the file after upload to the right filename back
+      rename(fileObj, fileObj.name, fileStoreStrategyFactory);
     },
     renameAttachment(fileObjId, newName) {
       check(fileObjId, String);
@@ -147,6 +150,56 @@ if (Meteor.isServer) {
 
       const fileObj = Attachments.findOne({_id: fileObjId});
       rename(fileObj, newName, fileStoreStrategyFactory);
+    },
+    validateAttachment(fileObjId) {
+      check(fileObjId, String);
+
+      const fileObj = Attachments.findOne({_id: fileObjId});
+      let isValid = true;
+
+      if (attachmentUploadMimeTypes.length) {
+        const mimeTypeResult = Promise.await(FileType.fromFile(fileObj.path));
+
+        const mimeType = (mimeTypeResult ? mimeTypeResult.mime : fileObj.type);
+        const baseMimeType = mimeType.split('/', 1)[0];
+
+        isValid = attachmentUploadMimeTypes.includes(mimeType) || attachmentUploadMimeTypes.includes(baseMimeType + '/*') || attachmentUploadMimeTypes.includes('*');
+
+        if (!isValid) {
+          console.log("Validation of uploaded file failed: file " + fileObj.path + " - mimetype " + mimeType);
+        }
+      }
+
+      if (attachmentUploadSize && fileObj.size > attachmentUploadSize) {
+        console.log("Validation of uploaded file failed: file " + fileObj.path + " - size " + fileObj.size);
+        isValid = false;
+      }
+
+      if (isValid && attachmentUploadExternalProgram) {
+        Promise.await(asyncExec(attachmentUploadExternalProgram.replace("{file}", '"' + fileObj.path + '"')));
+        isValid = fs.existsSync(fileObj.path);
+
+        if (!isValid) {
+          console.log("Validation of uploaded file failed: file " + fileObj.path + " has been deleted externally");
+        }
+      }
+
+      if (!isValid) {
+        Attachments.remove(fileObjId);
+      }
+    },
+    validateAttachmentAndMoveToStorage(fileObjId, storageDestination) {
+      check(fileObjId, String);
+      check(storageDestination, String);
+
+      Meteor.defer(() => Meteor.call('validateAttachment', fileObjId));
+
+      const fileObj = Attachments.findOne({_id: fileObjId});
+
+      if (fileObj) {
+        console.debug("Validation of uploaded file completed: file " + fileObj.path + " - storage destination " + storageDestination);
+        Meteor.defer(() => Meteor.call('moveAttachmentToStorage', fileObjId, storageDestination));
+      }
     },
   });